(feat) : QrCode 생성 로직 변경에 따른 로직 추가

+++ src/main/java/team/gwon/haveameal/ticket/mapper/TicketMapper.java
@@ -13,7 +13,7 @@ public interface TicketMapper {
 
 	List<PaymentWithCourseIncludeDetail> findAllMyTickets(MemberEntity member);
 
-	Integer getQrCode(PaymentWithCourseIncludeDetail paymentWithCourseIncludeDetail);
+	Long getQrCode(PaymentWithCourseIncludeDetail paymentWithCourseIncludeDetail);
 
 	Integer useQrCode(TicketEntity ticketEntity);
 }
